Fix service build on windows.
